Separated the crawl function
